Migration to Visual Studio 2019 (MSVC v142)
- Fix for building with Visual Studio 2019 (MSVC v142)
- Fix source file encoding
- Fix some simple bugs
- Replace with safe functions
- Reduce warnings
- Leave TODOs

+++ Video-Tools/AVIMux_GUI/AudioSource_AC3.cpp
@@ -316,7 +316,7 @@ int AC3SOURCE::ReadFrame(void* lpDest, DWORD* lpdwMicroSecRead,
 	double z=((double)dwRead/dwBitrate);
 
 	if (lpqwNanoSecRead) 
-		*lpqwNanoSecRead=round(8000000*z);
+		*lpqwNanoSecRead=(__int64)round(8000000*z);
 	
 	if (lpdwMicroSecRead) 
 		*lpdwMicroSecRead=(DWORD)round(8000*z);
